GUA-209: Rename files for consistency
The [Airbnb style guide](https://github.com/airbnb/javascript#naming--filename-matches-export)
says that the base filename should match the name of its default export.
None of these files have a default export, so we should fall back to
using camel case.

We're not strictly following the Airbnb style guide, but we should
make sure we're consistent with what we do and this seems like a simple
rule to follow.
