+++ src/lib/passportCheckVc.ts
@@ -5,23 +5,23 @@ import {
 
 import { RDF } from "@inrupt/vocab-common-rdf";
 
-import { 
+import {
+  CheckArtifacts,
   evidenceSuccessful,
   generateJWT,
   getBirthDate,
   getNameParts,
   GOV_UK_CREDENTIAL,
   GOV_UK_hasCredential
-} from "../lib/credential_helpers";
+} from "../lib/credentials";
 
-import { CheckArtifacts } from "../lib/credential_helpers";
 
 import {
   NamePart,
   PassportDetails,
-} from "../components/vocabularies/CommonComponents";
+} from "../components/vocabularies/commonComponents";
 
-import { Credential, IdentityCheck } from "../components/vocabularies/IdentityCheckCredential"
+import { Credential, IdentityCheck } from "../components/vocabularies/identityCheckCredential"
 
 // eslint-disable-next-line no-shadow
 import { Blob } from "node:buffer";
